requirements.txt
Django==3.0.4
django-filter==2.2.0
djangorestframework==3.11.0
djangorestframework-csv==2.1.0
geocoder==1.38.1
Markdown==3.2.1
psycopg2==2.8.4
